Split the packet-in/out declaration
This patch splits the declaration of the packet-in/out messages
from the echo-request/reply messages. As well, it provides a few
more unit tests to cover the serialization/deserialization of
the packet-in/out messages.
